# Optimize single NOT NULL count() aggregation