<!---
A technical comment, you are free to remove or leave it as it is when PR is created
The following categories are used in the next scripts, update them accordingly
utils/changelog/changelog.py
tests/ci/cancel_and_rerun_workflow_lambda/app.py
-->
### Changelog category (leave one):
- Performance Improvement


### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):

When the aggregation query contains only a single `COUNT()` function on a NOT NULL column, the aggregation logic is fully inlined during hash table probing. This avoids allocating and maintaining any aggregation state, significantly reducing memory usage and CPU overhead. This partially addresses #81982.

This optimization improves performance for queries like:

```sql
SELECT COUNT(*) FROM table GROUP BY key

SELECT COUNT(col) FROM table GROUP BY key  -- when `col` is NOT NULL
```

#### ClickBench Q15
```
HEAD:

Queries executed: 104.

localhost:9000, queries: 104, QPS: 4.881, RPS: 488124765.213, MiB/s: 3724.096, result RPS: 48.814, result MiB/s: 0.001.

0%              0.173 sec.
10%             0.184 sec.
20%             0.191 sec.
30%             0.193 sec.
40%             0.197 sec.
50%             0.200 sec.
60%             0.205 sec.
70%             0.210 sec.
80%             0.216 sec.
90%             0.220 sec.
95%             0.223 sec.
99%             0.228 sec.
99.9%           0.233 sec.
99.99%          0.233 sec.

This PR

Queries executed: 106.

localhost:9000, queries: 106, QPS: 6.301, RPS: 630035513.823, MiB/s: 4806.790, result RPS: 63.005, result MiB/s: 0.001.

0%              0.136 sec.
10%             0.143 sec.
20%             0.145 sec.
30%             0.150 sec.
40%             0.153 sec.
50%             0.156 sec.
60%             0.158 sec.
70%             0.159 sec.
80%             0.162 sec.
90%             0.166 sec.
95%             0.173 sec.
99%             0.179 sec.
99.9%           0.195 sec.
99.99%          0.195 sec.
```

### Documentation entry for user-facing changes

- [ ] Documentation is written (mandatory for new features)

<!---
Directly edit documentation source files in the "docs" folder with the same pull-request as code changes

or

Add a user-readable short description of the changes that should be added to docs.clickhouse.com below.

At a minimum, the following information should be added (but add more as needed).
- Motivation: Why is this function, table engine, etc. useful to ClickHouse users?

- Parameters: If the feature being added takes arguments, options or is influenced by settings, please list them below with a brief explanation.

- Example use: A query or command.
-->